\name{faultyscreen}
\alias{faultyscreen}
\docType{data}
\title{
faulty screen 
}
\description{
A screen with both systematic errors and sporadic contaminations. 
}
\usage{data(faultyscreen)}
\format{
  The format is:
 chr "cellHTS"
}

\examples{
data(faultyscreen)
## maybe str(faultyscreen) ; plot(faultyscreen) ...
}
\keyword{datasets}
